requirements.txt
fastapi==0.101.0
uvicorn==0.23.2
starlette==0.27.0
pydantic[email]==2.1.1
python-multipart==0.0.6
peewee==3.16.2
psycopg2-binary==2.9.7
pyjwt==2.8.0
fastapi-mail==1.4.1